ioutil: add WriteSerializeFile helper function
ioutil.WriteFile can lead to empty files if there's a crash or a disk
error (like out of space) at the wrong time.  Instead, overwrite and
truncate the existing file.  This has the same risk of partial OS
writes leading to corrupted files, without the risk of truncating a
non-empty file.

Issue: KBFS-2289
